+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1 - 2024-10-10]
+
 ### Fixed
 
 - Rimosse dipendenze non necessarie